backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
httpx==0.27.0
feedparser==6.0.11
sqlalchemy==2.0.30
asyncpg>=0.30.0
alembic==1.13.1
psycopg2-binary>=2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
redis==5.0.4
celery==5.4.0
qdrant-client==1.9.1
neo4j==5.20.0
langchain==0.2.5
langchain-openai==0.1.8
langchain-anthropic==0.1.15
langgraph==0.1.9
openai==1.30.0
anthropic==0.28.0
tiktoken>=0.8.0
sentence-transformers==3.0.0
python-docx==1.1.2
reportlab==4.2.0
pytesseract==0.3.10
Pillow>=11.0.0
pdf2image==1.17.0
pydantic>=2.7.4
pydantic-settings>=2.3.0
email-validator==2.1.1
tenacity==8.3.0
structlog==24.2.0
pytest==8.2.1
pytest-asyncio==0.23.7
factory-boy==3.3.0
